backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# TimescaleDB
sqlalchemy-timescaledb==0.1.0

# Data processing
pandas==2.1.4
numpy==1.26.3
lxml==5.1.0
python-dateutil==2.8.2

# Statistics
scipy==1.12.0

# Background jobs
celery==5.3.4
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Supabase
supabase==2.3.0

# File handling
aiofiles==23.2.1

# PDF generation
reportlab==4.0.9

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Email
aiosmtplib==3.0.1

# Monitoring
sentry-sdk[fastapi]==1.40.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
